[hwcomposer] Fix potential crash and slightly improve lock times for scheduleLayerList
The crash was that we didn't set acceptedLayerList to 0 when we started
another check. This would trigger the Q_ASSERT(acceptedLayerList==0)
further down.

The improvement is that we only need to mutex lock access to 'backend'
not the entire checkLayerList function. Once the m_scheduledLayerList
has been read on the HwcThread, we can release the mutex