[hwc] Do not power off display from HwComposerContext destructor. Fixes JB#34319
In SFOS mce should be in control of the logical display state. Having the
HwComposerContext objects power off display when destroyed can cause mce
to get out of sync with actual display state. That in turn can cause the
display to remain powered off for example when switching from pre user
session startup wizard to lipstick, or from lipstick to os update ui.