[hwcomposer] initialize binder thread pool, add fake SurfaceFlinger service. Fixes JB#41328
Initialize the binder thread pool by calling the startMiniSurfaceFlinger
function from libminisf.so (part of droidmedia) if available. This
avoids that binder services which are started from hwcomposer.*.so get
stuck forever.

The function also starts the fake SurfaceFlinger service if it is not
already started. This could improve performance on certain devices.