[hwcomposer] turn QPA_HWC_SYNC_BEFORE_SET into an environment variable. JB#41357
On some devices this helps to avoid non-causal screen transitions, i.e.
old frames being sporadically rendered at a later point in time again.
The issue is most noticeable in moving scenes.
It's hard to know which devices will be affected and how to detect
those, therefore changing this into an environment variable to have it
easily accessible.