[hwcomposer] Block in swap in the v11 backend too. Contributes to JB#45362
This is for slightly different reasons to the v20 backend where tearing
was visible in the rendering of internal scenes. Here we just want
to ensure the swap doesn't return until after a swap has actually
completed so we have a reference point where we know it's safe to
return wayland buffers to a client.