[hwcomposer] Block in swap waiting for the swap to actually complete. Contributes to JB#44212
The present fence should block rendering to a buffer until after a
swap and the acquire fence should block that same buffer from being
presented which accumulatively should prevent a rendering to a buffer
that is being presented. That's not happening on all hardware though
and we're seeing tearing, so we'll explicitly block in swap waiting
for the buffers from the previous swap to be released.