[hwcomposer] Separate legacy hwc quirks and do not use them with hwc2. JB#62779 JB#62780
In legacy adaptations we start a dummy SurfaceFlinger service via libminisf and
drop the minisf service from droidmedia. This has two reasons:

1. The android binder thread pool gets initialised which means that functionality
   that uses binder within the hwcomposer module works properly. e.g. dumpsys does
   not get stuck on display.qservice and other issues such as on certain devices
   media doesn't get stuck attempting to communicate with services in hwcomposer.
2. It may improve performance in weird hwcomposer implementations which attempt to
   communicate with SurfaceFlinger because binder communication can be handled
   without leaving the process.

In hwc2 all this is not needed anymore as the hwc2 module is part of a separate
composer service. This loading libminisf also causes issues with latest Android
14 base.